Added basic axum server Restructured the whole thing 'cause I couldn't deal with it
