+++ mocker/Cargo.toml
@@ -17,3 +17,5 @@ serde_json = "1.0.110"
 starduck = "0.1"
 uuid = { version = "1.6.1", features = ["serde", "v4"] }
 openssl = { version = "0.10.59", features = ["vendored"] }
+axum = "0.7.4"
+tokio = { version = "1.35.1", features = ["full"] }
