+++ mocker/src/message/mod.rs
@@ -1,0 +1,10 @@
+mod builders;
+mod messenger;
+mod processors;
+
+pub use builders::{build_duration, build_message};
+pub use messenger::messenger;
+pub use processors::process_args;
+
+use builders::{build_mqtt_client, build_tuples};
+use processors::{process_fixed_values, process_random_values};
